requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14,<3.16
drf-spectacular>=0.27,<0.28
Pillow>=10.0
